core: change page idx type from usize to i64
MVCC is like the annoying younger cousin (I know because I was him) that
needs to be treated differently. MVCC requires us to use root_pages that
might not be allocated yet, and the plan is to use negative root_pages
for that case. Therefore, we need i64 in order to fit this change.

+++ core/error.rs
@@ -123,12 +123,12 @@ pub enum CompletionError {
     #[error("Completion was aborted")]
     Aborted,
     #[error("Decryption failed for page={page_idx}")]
-    DecryptionError { page_idx: usize },
+    DecryptionError { page_idx: i64 },
     #[error("I/O error: partial write")]
     ShortWrite,
     #[error("Checksum mismatch on page {page_id}: expected {expected}, got {actual}")]
     ChecksumMismatch {
-        page_id: usize,
+        page_id: i64,
         expected: u64,
         actual: u64,
     },
